chore(android): move utility methods to new class
Refactor charSequence utilities to CharSequenceUtil.java
