+++ android/KMEA/app/src/main/java/com/tavultesoft/kmea/KMManager.java
@@ -61,6 +61,7 @@ import com.tavultesoft.kmea.packages.JSONUtils;
 import com.tavultesoft.kmea.packages.LexicalModelPackageProcessor;
 import com.tavultesoft.kmea.packages.PackageProcessor;
 import com.tavultesoft.kmea.KMScanCodeMap;
+import com.tavultesoft.kmea.util.CharSequenceUtil;
 import com.tavultesoft.kmea.util.FileUtils;
 
 import org.json.JSONArray;
@@ -309,94 +310,6 @@ public final class KMManager {
     }
   }
 
-  // These aren't public for KMManager API. Just for unit testing:
-
-  /**
-   * Count the number of surrogate pairs starting from the end of a character sequence
-   * until we reach dn codepoints or searched the entire sequence.
-   * Doing this the hard way because we can't foreach charSequence in reverse.
-   * @param sequence - the character sequence to analyze
-   * @param dn - the number of code points to count up to
-   * @return int
-   */
-  public static int countSurrogatePairs(CharSequence sequence, int dn) {
-    if ((sequence == null) || (sequence.length() == 0) || (dn <= 0)) {
-      return 0;
-    }
-
-    int numPairs = 0, counter = 0;
-    int lastIndex = sequence.length()-1;
-    try {
-      do {
-        if (Character.isLowSurrogate(sequence.charAt(lastIndex - counter))) {
-          numPairs++;
-        }
-        counter++;
-      } while ((lastIndex - counter >= 0) && (counter < dn) && (numPairs < dn));
-      return numPairs;
-    } catch (Exception e) {
-      Log.e(TAG, "Error in countSurrogatePairs: " + e);
-      return numPairs;
-    }
-  }
-
-  /**
-   * Determine a character sequence that needs to be re-inserted. If currentContext is a
-   * subSequence of expectedChars, returns the character sequence that needs to be restored.
-   * @param expectedChars - expected character sequence
-   * @param currentContext - current character sequence
-   * @return CharSequence - Character sequence that will need to be appended to currentContext.
-   *                        Empty string if no characters need to be restored.
-   */
-  public static CharSequence restoreChars(CharSequence expectedChars, CharSequence currentContext) {
-    CharSequence charsToRestore = "";
-
-    try {
-      // Now see if we need to re-insert characters
-      if (expectedChars.length() != currentContext.length()) {
-        String expectedCharsString = expectedChars.toString();
-        String currentContextString = currentContext.toString();
-        int index = expectedCharsString.indexOf(currentContextString);
-        if (index > -1) {
-          // subSequence indices are start(inclusive) to end(exclusive)
-          charsToRestore = expectedChars.subSequence(index + currentContextString.length(), expectedChars.length());
-        }
-      }
-
-      return charsToRestore;
-    } catch (Exception e) {
-      Log.e(TAG, "Error in restoreChars: " + e);
-      return charsToRestore;
-    }
-  }
-
-  /**
-   * Adjust cursor position after insert the new text.
-   * @param charsBefore the character sequence of the inputconnection (2*s.length)
-   * @param s the inserted text
-   * @return int - number of characters to adjust the cursor
-   */
-  public static int adjustCursorPosition(CharSequence charsBefore, String s) {
-    if (charsBefore == null || s == null) {
-      return 0;
-    }
-
-    int _expected_start_index = charsBefore.length() - s.length();
-    int _move = 0;
-    while (_move < _expected_start_index) {
-
-      CharSequence _check = charsBefore.subSequence(
-        _expected_start_index - _move,charsBefore.length()-_move);
-      if(_check.equals(s)) {
-        break;
-      }
-      _move++;
-    }
-    return _move;
-  }
-
-  // End of public methods for unit testing
-
   /*
   // TODO: Chromium has a bug where deleteSurroundingText deletes an entire grapheme cluster
   // instead of one code-point.
@@ -423,7 +336,7 @@ public final class KMManager {
       return;
     }
 
-    int numPairs = countSurrogatePairs(charsBackup, dn);
+    int numPairs = CharSequenceUtil.countSurrogatePairs(charsBackup, dn);
 
     // Chop dn+numPairs code points from the end of charsBackup
     // subSequence indices are start(inclusive) to end(exclusive)
@@ -431,7 +344,7 @@ public final class KMManager {
     ic.deleteSurroundingText(dn + numPairs, 0);
     CharSequence newContext = ic.getTextBeforeCursor(originalBufferLength - dn, 0);
 
-    CharSequence charsToRestore = restoreChars(expectedChars, newContext);
+    CharSequence charsToRestore = CharSequenceUtil.restoreChars(expectedChars, newContext);
     if (charsToRestore.length() > 0) {
       // Restore expectedChars that Chromium deleted, and advance the cursor by expectedChars.length()
       ic.commitText(charsToRestore, charsToRestore.length());
@@ -2206,7 +2119,7 @@ public final class KMManager {
 
             //String userAgent = System.getProperty("http.agent");
             CharSequence charsBefore = ic.getTextBeforeCursor(s.length()*2, 0);
-            int move = adjustCursorPosition(charsBefore, s);
+            int move = CharSequenceUtil.adjustCursorPosition(charsBefore, s);
             if (move > 0) {
               Log.d(TAG, "adjusting cursor charsBefore: " + charsBefore.toString() + ", s: " + s + ", move: " + move);
               ic.commitText("", -move);
